Can round trip multiple simultaneous streams into the BlockMultiStream structure,

+++ src/Melville.FileSystem/BlockFile/BlockMultiStreams/MemoryOperations.cs
@@ -4,9 +4,9 @@ namespace Melville.FileSystem.BlockFile.BlockMultiStreams;
 
 public static class MemoryOperations
 {
-    public static Memory<T> OfMaxLen<T>(this Memory<T> memory, long length) => 
+    public static Memory<T> OfMaxLen<T>(this Memory<T> memory, int length) => 
         length < memory.Length ? memory.Slice(0, (int)length) : memory;
 
-    public static ReadOnlyMemory<T> OfMaxLen<T>(this ReadOnlyMemory<T> memory, long length) => 
+    public static ReadOnlyMemory<T> OfMaxLen<T>(this ReadOnlyMemory<T> memory, int length) => 
         length < memory.Length ? memory.Slice(0, (int)length) : memory;
 }
